Implement PT path control
Now we can make PT render any specified path types, like LS*E or {LSE and LDE}

+++ src/OptiXPathTracer/whitted.h
@@ -99,7 +99,7 @@ struct PayloadRadiance
     unsigned int seed;
     bool done;
     bool glossy_bounce;
-    short caustic_bounce_state;
+    long long caustic_bounce_state;
     RT_FUNCTION PayloadRadiance()
     {
         result = make_float3(0);
